# docs/api-docs-horseOwner-Trainer-controller.yaml
openapi: 3.1.0
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:8080
  description: Generated server url
tags:
- name: Trainer
  description: Trainer management APIs
- name: Authentication
  description: Authentication management APIs
- name: Horse Owner
  description: Horse Owner management APIs
paths:
  /api/trainer/complete-profile:
    put:
      tags:
      - Trainer
      operationId: completeProfile
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompleteTrainerProfileRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseTrainerProfileResponse"
  /api/horse-owner/horses/{horseId}/assign-trainer:
    put:
      tags:
      - Horse Owner
      operationId: assignTrainer
      parameters:
      - name: horseId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: trainerId
        in: query
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseSignHorseResponse"
  /api/admin/users/{id}/status:
    put:
      tags:
      - admin-controller
      operationId: updateStatus
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserStatusRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
  /api/admin/users/{id}/role:
    put:
      tags:
      - admin-controller
      operationId: updateRole
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRoleRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
  /api/horse-owner/horses:
    get:
      tags:
      - Horse Owner
      operationId: getHorseList
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseListSignHorseResponse"
    post:
      tags:
      - Horse Owner
      operationId: signHorse
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignHorseRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseSignHorseResponse"
  /api/auth/verify-reset-otp:
    post:
      tags:
      - Authentication
      operationId: verifyResetOtp
      parameters:
      - name: email
        in: query
        required: true
        schema:
          type: string
      - name: otp
        in: query
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: string
  /api/auth/verify-email:
    post:
      tags:
      - Authentication
      operationId: verifyEmail
      parameters:
      - name: email
        in: query
        required: true
        schema:
          type: string
      - name: otp
        in: query
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: string
  /api/auth/send-verification-otp:
    post:
      tags:
      - Authentication
      operationId: sendVerificationOtp
      parameters:
      - name: email
        in: query
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: string
  /api/auth/reset-password:
    post:
      tags:
      - Authentication
      operationId: resetPassword
      parameters:
      - name: otp
        in: query
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetPasswordRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseAuthMeResponse"
  /api/auth/register:
    post:
      tags:
      - Authentication
      summary: Register a new user account
      operationId: register
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseAuthMeResponse"
  /api/auth/logout:
    post:
      tags:
      - Authentication
      operationId: logout
      parameters:
      - name: Authorization
        in: header
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseString"
  /api/auth/login:
    post:
      tags:
      - Authentication
      operationId: login
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseLoginResponse"
  /api/auth/forgot-password:
    post:
      tags:
      - Authentication
      operationId: forgotPassword
      parameters:
      - name: email
        in: query
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: string
  /api/trainer/profile:
    get:
      tags:
      - Trainer
      operationId: getProfile
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseTrainerProfileResponse"
  /api/horse-owner/horses/{horseId}:
    get:
      tags:
      - Horse Owner
      operationId: getHorse
      parameters:
      - name: horseId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseSignHorseResponse"
  /api/auth/me:
    get:
      tags:
      - Authentication
      summary: Get current authenticated user profile
      operationId: me
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseAuthMeResponse"
  /api/admin/users:
    get:
      tags:
      - admin-controller
      operationId: getUsers
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      - name: size
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 10
      - name: keyword
        in: query
        required: false
        schema:
          type: string
      - name: role
        in: query
        required: false
        schema:
          type: string
          enum:
          - ADMIN
          - STAFF
          - SPECTATOR
          - USER
          - HORSE_OWNER
          - JOCKEY
          - REFEREE
          - MANAGER
          - TRAINER
      - name: status
        in: query
        required: false
        schema:
          type: string
          enum:
          - ACTIVE
          - INACTIVE
          - BANNED
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponsePageResponseAdminUserItemResponse"
components:
  schemas:
    CompleteTrainerProfileRequest:
      type: object
      properties:
        age:
          type: integer
          format: int32
        experienceYears:
          type: integer
          format: int32
        description:
          type: string
        avatarUrl:
          type: string
    ApiResponseTrainerProfileResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          $ref: "#/components/schemas/TrainerProfileResponse"
    TrainerProfileResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        userId:
          type: integer
          format: int64
        name:
          type: string
        age:
          type: integer
          format: int32
        experienceYears:
          type: integer
          format: int32
        description:
          type: string
        avatarUrl:
          type: string
        status:
          type: string
    ApiResponseSignHorseResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          $ref: "#/components/schemas/SignHorseResponse"
    SignHorseResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        horseName:
          type: string
        breed:
          type: string
        age:
          type: integer
          format: int32
        gender:
          type: string
        speedRating:
          type: integer
          format: int32
        historyRank:
          type: string
        avatarUrl:
          type: string
        weight:
          type: integer
          format: int64
        status:
          type: string
          enum:
          - ACTIVE
          - INACTIVE
          - RETIRE
        trainerName:
          type: string
        trainerId:
          type: integer
          format: int64
        ownerId:
          type: integer
          format: int64
        ownerName:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    UpdateUserStatusRequest:
      type: object
      properties:
        status:
          type: string
          enum:
          - ACTIVE
          - INACTIVE
          - BANNED
      required:
      - status
    ApiResponseVoid:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data: {}
    UpdateUserRoleRequest:
      type: object
      properties:
        roleName:
          type: string
          enum:
          - ADMIN
          - STAFF
          - SPECTATOR
          - USER
          - HORSE_OWNER
          - JOCKEY
          - REFEREE
          - MANAGER
          - TRAINER
      required:
      - roleName
    SignHorseRequest:
      type: object
      properties:
        horseName:
          type: string
        breed:
          type: string
        age:
          type: integer
          format: int32
        gender:
          type: string
        speedRating:
          type: integer
          format: int32
        history_rank:
          type: string
        avatar_url:
          type: string
        weight:
          type: integer
          format: int64
        status:
          type: string
          enum:
          - ACTIVE
          - INACTIVE
          - RETIRE
    ResetPasswordRequest:
      type: object
      properties:
        email:
          type: string
          minLength: 1
        newPassWord:
          type: string
          maxLength: 2147483647
          minLength: 8
      required:
      - email
      - newPassWord
    ApiResponseAuthMeResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          $ref: "#/components/schemas/AuthMeResponse"
    AuthMeResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        username:
          type: string
        password:
          type: string
        role:
          type: string
        email:
          type: string
        fullName:
          type: string
        verified:
          type: boolean
        phoneNumber:
          type: string
        avatar:
          type: string
        status:
          type: string
    RegisterRequest:
      type: object
      properties:
        role:
          type: string
          enum:
          - ADMIN
          - STAFF
          - SPECTATOR
          - USER
          - HORSE_OWNER
          - JOCKEY
          - REFEREE
          - MANAGER
          - TRAINER
        fullName:
          type: string
          minLength: 1
        username:
          type: string
          minLength: 1
        email:
          type: string
          minLength: 1
        password:
          type: string
          maxLength: 2147483647
          minLength: 8
        phone:
          type: string
          pattern: "^[0-9]{10,11}$"
      required:
      - email
      - fullName
      - password
      - username
    ApiResponseString:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          type: string
    LoginRequest:
      type: object
      properties:
        username:
          type: string
          minLength: 1
        password:
          type: string
          minLength: 1
      required:
      - password
      - username
    ApiResponseLoginResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          $ref: "#/components/schemas/LoginResponse"
    LoginResponse:
      type: object
      properties:
        accessToken:
          type: string
        tokenType:
          type: string
        expiresIn:
          type: integer
          format: int64
        user:
          $ref: "#/components/schemas/AuthMeResponse"
    ApiResponseListSignHorseResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          type: array
          items:
            $ref: "#/components/schemas/SignHorseResponse"
    AdminUserItemResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        fullName:
          type: string
        email:
          type: string
        phone:
          type: string
        role:
          type: string
        status:
          type: string
        createdAt:
          type: string
          format: date-time
    ApiResponsePageResponseAdminUserItemResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          $ref: "#/components/schemas/PageResponseAdminUserItemResponse"
    PageResponseAdminUserItemResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/AdminUserItemResponse"
        page:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
        first:
          type: boolean
        last:
          type: boolean